Convert signalR to a function component
`SignalRConnector.js` becomes `SignalRListener.tsx`, following Sonarr's
c3f9cd12. Redux stays inside -- every dispatch is still a dispatch, just via
`useDispatch` -- so this is a shape change, not a migration. It is done now so
Queue and Collection land in a converted file instead of forcing a second
rewrite.

The dynamic `this[getHandlerName(name)]` lookup could not survive the move to a
function component, so handlers become an explicit name chain as upstream has
it. The unknown-message `console.error` is kept; Sonarr dropped it.

One subtle hazard worth naming. Handlers live in refs so the connection is
built once, which means anything they close over is frozen at mount. That is
fine for `dispatch`, which is stable, but `handleQueue` reads
`state.queue.paged.isPopulated` at message time -- captured in a ref it would
freeze at its initial `false` and queue refreshes would silently stop once the
queue populated. It is held in a ref updated every render instead, preserving
the class's read-at-call-time semantics.

The signalR import moves from `@microsoft/signalr/dist/browser/signalr.js` to
the package root. The deep path ships no types, so under TypeScript it was an
implicit `any` that failed the webpack build while `tsc` passed; the root
resolves `dist/esm` with declarations. Sonarr imports the same way.

`Logger` is extracted to `Utilities/SignalRLogger.ts`, also as upstream.
