+++ frontend/src/Components/Page/Page.tsx
@@ -5,7 +5,7 @@ import ColorImpairedContext from 'App/ColorImpairedContext';
 import ConnectionLostModal from 'App/ConnectionLostModal';
 import AppState from 'App/State/AppState';
 import { SafeForWorkModeContext } from 'App/State/SafeForWorkContext';
-import SignalRConnector from 'Components/SignalRConnector';
+import SignalRListener from 'Components/SignalRListener';
 import AuthenticationRequiredModal from 'FirstRun/AuthenticationRequiredModal';
 import useAppPage from 'Helpers/Hooks/useAppPage';
 import { saveDimensions } from 'Store/Actions/appActions';
@@ -92,7 +92,7 @@ function Page({ children }: PageProps) {
     <SafeForWorkModeContext.Provider value={safeForWorkMode}>
       <ColorImpairedContext.Provider value={enableColorImpairedMode}>
         <div className={styles.page}>
-          <SignalRConnector />
+          <SignalRListener />
 
           <PageHeader isSmallScreen={isSmallScreen} />
