--- frontend/src/Components/SignalRConnector.js
@@ -1,654 +1,0 @@
-// TODO: Standardize the "when" on React Query Key invalidation to avoid rapid-fire reloads during bulk operations.
-import * as signalR from '@microsoft/signalr/dist/browser/signalr.js';
-import PropTypes from 'prop-types';
-import { Component } from 'react';
-import { connect } from 'react-redux';
-import { createSelector } from 'reselect';
-import { queryClient } from 'App/queryClient';
-import { setAppValue, setVersion } from 'Store/Actions/appActions';
-import {
-  removeItem,
-  update,
-  updateItem,
-  updateItemsBatch,
-} from 'Store/Actions/baseActions';
-import {
-  fetchCommands,
-  finishCommand,
-  updateCommand,
-} from 'Store/Actions/commandActions';
-import { fetchQueue, fetchQueueDetails } from 'Store/Actions/queueActions';
-import { fetchRootFolders } from 'Store/Actions/rootFolderActions';
-import { fetchQualityDefinitions } from 'Store/Actions/settingsActions';
-import { fetchTagDetails, fetchTags } from 'Store/Actions/tagActions';
-import { repopulatePage } from 'Utilities/pagePopulator';
-import titleCase from 'Utilities/String/titleCase';
-
-function getHandlerName(name) {
-  name = titleCase(name);
-  name = name.replace('/', '');
-
-  return `handle${name}`;
-}
-
-function createMapStateToProps() {
-  return createSelector(
-    (state) => state.app.isReconnecting,
-    (state) => state.app.isDisconnected,
-    (state) => state.queue.paged.isPopulated,
-    (isReconnecting, isDisconnected, isQueuePopulated) => {
-      return {
-        isReconnecting,
-        isDisconnected,
-        isQueuePopulated,
-      };
-    }
-  );
-}
-
-const mapDispatchToProps = {
-  dispatchFetchCommands: fetchCommands,
-  dispatchUpdateCommand: updateCommand,
-  dispatchFinishCommand: finishCommand,
-  dispatchSetAppValue: setAppValue,
-  dispatchSetVersion: setVersion,
-  dispatchUpdate: update,
-  dispatchUpdateItem: updateItem,
-  dispatchUpdateItemsBatch: updateItemsBatch,
-  dispatchRemoveItem: removeItem,
-  dispatchFetchQualityDefinitions: fetchQualityDefinitions,
-  dispatchFetchQueue: fetchQueue,
-  dispatchFetchQueueDetails: fetchQueueDetails,
-  dispatchFetchRootFolders: fetchRootFolders,
-  dispatchFetchTags: fetchTags,
-  dispatchFetchTagDetails: fetchTagDetails,
-};
-
-function Logger(minimumLogLevel) {
-  this.minimumLogLevel = minimumLogLevel;
-}
-
-Logger.prototype.cleanse = function (message) {
-  const apikey = new RegExp(
-    `access_token=${encodeURIComponent(window.Whisparr.apiKey)}`,
-    'g'
-  );
-  return message.replace(apikey, 'access_token=(removed)');
-};
-
-Logger.prototype.log = function (logLevel, message) {
-  // see https://github.com/aspnet/AspNetCore/blob/21c9e2cc954c10719878839cd3f766aca5f57b34/src/SignalR/clients/ts/signalr/src/Utils.ts#L147
-  if (logLevel >= this.minimumLogLevel) {
-    switch (logLevel) {
-      case signalR.LogLevel.Critical:
-      case signalR.LogLevel.Error:
-        console.error(
-          `[signalR] ${signalR.LogLevel[logLevel]}: ${this.cleanse(message)}`
-        );
-        break;
-      case signalR.LogLevel.Warning:
-        console.warn(
-          `[signalR] ${signalR.LogLevel[logLevel]}: ${this.cleanse(message)}`
-        );
-        break;
-      case signalR.LogLevel.Information:
-        console.info(
-          `[signalR] ${signalR.LogLevel[logLevel]}: ${this.cleanse(message)}`
-        );
-        break;
-      default:
-        // console.debug only goes to attached debuggers in Node, so we use console.log for Trace and Debug
-        console.log(
-          `[signalR] ${signalR.LogLevel[logLevel]}: ${this.cleanse(message)}`
-        );
-        break;
-    }
-  }
-};
-
-// Helper to ensure updated data when a movie is updated in the background
-function updateMovieDetailsQueryCache(updatedMovie) {
-  const queryKey = `/movie/${updatedMovie.titleSlug}`;
-
-  // Don't trigger a re-fetch, as this method can be called rapid-fire
-  queryClient.invalidateQueries({
-    queryKey: [queryKey],
-    refetchType: 'none',
-  });
-}
-
-// Helper to update a nested movie in performer.years[].movies[] in React Query cache
-// Update Performer React Query cache helper
-function updateMovieInPerformerWorksQueryCache(updatedMovie) {
-  // Find all queries for performer works
-  const queryCache = queryClient.getQueryCache().findAll();
-  queryCache.forEach(({ queryKey }) => {
-    // Look for keys like "/performer/{performerId}/works"
-    if (
-      Array.isArray(queryKey) &&
-      typeof queryKey[0] === 'string' &&
-      /performer\/[^/]+\/works$/.test(queryKey[0])
-    ) {
-      queryClient.setQueryData(queryKey, (oldData) => {
-        if (!Array.isArray(oldData)) {
-          return oldData;
-        }
-        const idx = oldData.findIndex((movie) => movie.id === updatedMovie.id);
-        if (idx === -1) {
-          return oldData;
-        }
-        const newData = [...oldData];
-        newData[idx] = { ...oldData[idx], ...updatedMovie };
-        return newData;
-      });
-    }
-  });
-}
-
-// Helper to update a nested movie in performer.years[].movies[] in React Query cache
-// To avoid invalidating the entire list when a single movie is updated, we need to find the movie in the cache and update it directly.
-function updateMovieInStudioWorksQueryCache(updatedMovie) {
-  if (!updatedMovie?.studioForeignId) {
-    return;
-  }
-
-  const studioKey = `/studio/${updatedMovie.studioForeignId}/works`;
-  queryClient.setQueryData([studioKey], (oldData) => {
-    if (!Array.isArray(oldData)) {
-      return oldData;
-    }
-    const idx = oldData.findIndex((movie) => movie.id === updatedMovie.id);
-    if (idx === -1) {
-      return oldData;
-    }
-    const newData = [...oldData];
-    newData[idx] = { ...oldData[idx], ...updatedMovie };
-    return newData;
-  });
-}
-
-// Merges updates in React Queryr cache instead of a re-fetch
-function updatePerformerQueryCache(updatedPerformer) {
-  if (!updatedPerformer?.foreignId) {
-    return;
-  }
-
-  const performerKey = `/performer/${updatedPerformer.foreignId}`;
-
-  queryClient.setQueryData([performerKey], (oldData) => {
-    if (!oldData || typeof oldData !== 'object') {
-      return updatedPerformer;
-    }
-    return { ...oldData, ...updatedPerformer };
-  });
-}
-
-// Merges updates in React Queryr cache instead of a re-fetch
-function updateStudioQueryCache(updatedStudio) {
-  if (!updatedStudio?.foreignId) {
-    return;
-  }
-
-  const studioKey = `/studio/${updatedStudio.foreignId}`;
-  queryClient.setQueryData([studioKey], (oldData) => {
-    if (!oldData || typeof oldData !== 'object') {
-      return updatedStudio;
-    }
-    return { ...oldData, ...updatedStudio };
-  });
-}
-
-function removeMovieQueryCache(updatedMovie) {
-  if (!updatedMovie?.foreignId) {
-    return;
-  }
-
-  queryClient.removeQueries({
-    queryKey: [`/movie/${updatedMovie.titleSlug}`],
-  });
-}
-
-function removePerformerQueryCache(updatedPerformer) {
-  if (!updatedPerformer?.foreignId) {
-    return;
-  }
-
-  queryClient.removeQueries({
-    queryKey: [`/performer/${updatedPerformer.foreignId}`],
-  });
-
-  queryClient.removeQueries({
-    queryKey: [`/performer/${updatedPerformer.foreignId}/works`],
-  });
-}
-
-function removeStudioQueryCache(updatedStudio) {
-  if (!updatedStudio?.foreignId) {
-    return;
-  }
-
-  queryClient.removeQueries({
-    queryKey: [`/studio/${updatedStudio.foreignId}`],
-  });
-
-  queryClient.removeQueries({
-    queryKey: [`/studio/${updatedStudio.foreignId}/works`],
-  });
-}
-
-function invalidateMoviePagedQueryCache() {
-  queryClient.invalidateQueries({
-    queryKey: ['/movie/paged'],
-  });
-}
-
-function invalidatePerformerPagedQueryCache() {
-  queryClient.invalidateQueries({
-    predicate: (query) => {
-      return (
-        Array.isArray(query.queryKey) &&
-        typeof query.queryKey[0] === 'string' &&
-        query.queryKey[0].startsWith('/performer/paged')
-      );
-    },
-  });
-}
-
-function invalidateStudioPagedQueryCache() {
-  queryClient.invalidateQueries({
-    predicate: (query) => {
-      return (
-        Array.isArray(query.queryKey) &&
-        typeof query.queryKey[0] === 'string' &&
-        query.queryKey[0].startsWith('/studio/paged')
-      );
-    },
-  });
-}
-
-class SignalRConnector extends Component {
-  //
-  // Lifecycle
-
-  constructor(props, context) {
-    super(props, context);
-
-    this.connection = null;
-  }
-
-  componentDidMount() {
-    console.log('[signalR] starting');
-
-    const url = `${window.Whisparr.urlBase}/signalr/messages`;
-
-    this.connection = new signalR.HubConnectionBuilder()
-      .configureLogging(new Logger(signalR.LogLevel.Information))
-      .withUrl(
-        `${url}?access_token=${encodeURIComponent(window.Whisparr.apiKey)}`
-      )
-      .withAutomaticReconnect({
-        nextRetryDelayInMilliseconds: (retryContext) => {
-          if (retryContext.elapsedMilliseconds > 180000) {
-            this.props.dispatchSetAppValue({ isDisconnected: true });
-          }
-          return Math.min(retryContext.previousRetryCount, 10) * 1000;
-        },
-      })
-      .build();
-
-    this.connection.onreconnecting(this.onReconnecting);
-    this.connection.onreconnected(this.onReconnected);
-    this.connection.onclose(this.onClose);
-
-    this.connection.on('receiveMessage', this.onReceiveMessage);
-
-    this.connection.start().then(this.onStart, this.onStartFail);
-  }
-
-  componentWillUnmount() {
-    this.connection.stop();
-    this.connection = null;
-  }
-
-  //
-  // Control
-  handleMessage = (message) => {
-    const { name, body } = message;
-
-    const handler = this[getHandlerName(name)];
-
-    if (handler) {
-      handler(body);
-      return;
-    }
-
-    console.error(`signalR: Unable to find handler for ${name}`);
-  };
-
-  handleCalendar = (body) => {
-    if (body.action === 'updated') {
-      this.props.dispatchUpdateItem({
-        section: 'calendar',
-        updateOnly: true,
-        ...body.resource,
-      });
-    }
-  };
-
-  handleCommand = (body) => {
-    if (body.action === 'sync') {
-      this.props.dispatchFetchCommands();
-      return;
-    }
-
-    const resource = body.resource;
-    const status = resource.status;
-
-    // Both successful and failed commands need to be
-    // completed, otherwise they spin until they time out.
-
-    if (status === 'completed' || status === 'failed') {
-      this.props.dispatchFinishCommand(resource);
-    } else {
-      this.props.dispatchUpdateCommand(resource);
-    }
-  };
-
-  handleMoviefile = (body) => {
-    if (body.action === 'updated' || body.action === 'deleted') {
-      queryClient.invalidateQueries({ queryKey: ['/moviefile'] });
-    }
-  };
-
-  handleDownloadclient = ({ action, resource }) => {
-    const section = 'settings.downloadClients';
-
-    if (action === 'created' || action === 'updated') {
-      this.props.dispatchUpdateItem({ section, ...resource });
-    } else if (action === 'deleted') {
-      this.props.dispatchRemoveItem({ section, id: resource.id });
-    }
-  };
-
-  handleHealth = () => {
-    queryClient.invalidateQueries({ queryKey: ['/health'] });
-  };
-
-  handleImportlist = ({ action, resource }) => {
-    const section = 'settings.importLists';
-
-    if (action === 'created' || action === 'updated') {
-      this.props.dispatchUpdateItem({ section, ...resource });
-    } else if (action === 'deleted') {
-      this.props.dispatchRemoveItem({ section, id: resource.id });
-    }
-  };
-
-  handleIndexer = ({ action, resource }) => {
-    const section = 'settings.indexers';
-
-    if (action === 'created' || action === 'updated') {
-      this.props.dispatchUpdateItem({ section, ...resource });
-    } else if (action === 'deleted') {
-      this.props.dispatchRemoveItem({ section, id: resource.id });
-    }
-  };
-
-  handleNotification = ({ action, resource }) => {
-    const section = 'settings.notifications';
-
-    if (action === 'created' || action === 'updated') {
-      this.props.dispatchUpdateItem({ section, ...resource });
-    } else if (action === 'deleted') {
-      this.props.dispatchRemoveItem({ section, id: resource.id });
-    }
-  };
-
-  handleMovie = (body) => {
-    // Support batch payloads (Resources) and single (resource)
-    if (Array.isArray(body.resources) && body.resources.length > 0) {
-      // Batched update
-      if (body.action === 'updated') {
-        body.resources.forEach(updateMovieInPerformerWorksQueryCache);
-        body.resources.forEach(updateMovieInStudioWorksQueryCache);
-        body.resources.forEach(updateMovieDetailsQueryCache);
-      } else if (body.action === 'deleted') {
-        body.resources.forEach(removeMovieQueryCache);
-      }
-
-      invalidateMoviePagedQueryCache();
-      return;
-    }
-
-    // Fallback: single resource
-    const action = body.action;
-    if (action === 'updated') {
-      updateMovieInPerformerWorksQueryCache(body.resource);
-      updateMovieInStudioWorksQueryCache(body.resource);
-      updateMovieDetailsQueryCache(body.resource);
-    } else if (action === 'deleted') {
-      removeMovieQueryCache(body.resource);
-    }
-    invalidateMoviePagedQueryCache();
-  };
-
-  handleCollection = (body) => {
-    const action = body.action;
-    const section = 'movieCollections';
-
-    console.log(body);
-
-    if (action === 'updated') {
-      this.props.dispatchUpdateItem({ section, ...body.resource });
-    } else if (action === 'deleted') {
-      this.props.dispatchRemoveItem({ section, id: body.resource.id });
-    }
-  };
-
-  handlePerformer = (body) => {
-    const action = body.action;
-
-    if (Array.isArray(body.resources) && body.resources.length > 0) {
-      if (action === 'deleted') {
-        body.resources.forEach((resource) => {
-          removePerformerQueryCache(resource);
-        });
-      } else {
-        body.resources.forEach((resource) => {
-          updatePerformerQueryCache(resource);
-        });
-      }
-      invalidatePerformerPagedQueryCache();
-      return;
-    }
-
-    if (action === 'updated') {
-      // Update individual performer query keys rather than re-fetch
-      updatePerformerQueryCache(body.resource);
-      // Force paged query re-fetch so updates are immediate
-      invalidatePerformerPagedQueryCache();
-    } else if (action === 'deleted') {
-      // Remove individual performer quey keys
-      removePerformerQueryCache(body.resource);
-      // Force paged query re-fetch so updates are immediate
-      invalidatePerformerPagedQueryCache();
-    }
-  };
-
-  handleQualitydefinition = () => {
-    this.props.dispatchFetchQualityDefinitions();
-  };
-
-  handleQueue = () => {
-    if (this.props.isQueuePopulated) {
-      this.props.dispatchFetchQueue();
-    }
-  };
-
-  handleQueueDetails = () => {
-    this.props.dispatchFetchQueueDetails();
-  };
-
-  handleQueueStatus = (body) => {
-    this.props.dispatchUpdate({ section: 'queue.status', data: body.resource });
-  };
-
-  handleStudio = (body) => {
-    const action = body.action;
-
-    if (Array.isArray(body.resources) && body.resources.length > 0) {
-      if (action === 'deleted') {
-        body.resources.forEach((resource) => {
-          removeStudioQueryCache(resource);
-        });
-      } else {
-        body.resources.forEach((resource) => {
-          updateStudioQueryCache(resource);
-        });
-      }
-      invalidateStudioPagedQueryCache();
-      return;
-    }
-
-    if (action === 'updated') {
-      updateStudioQueryCache(body.resource);
-      invalidateStudioPagedQueryCache();
-    } else if (action === 'deleted') {
-      removeStudioQueryCache(body.resource);
-      invalidateStudioPagedQueryCache();
-    }
-  };
-
-  handleVersion = (body) => {
-    const version = body.version;
-
-    this.props.dispatchSetVersion({ version });
-  };
-
-  handleWantedCutoff = (body) => {
-    if (body.action === 'updated') {
-      this.props.dispatchUpdateItem({
-        section: 'wanted.cutoffUnmet',
-        updateOnly: true,
-        ...body.resource,
-      });
-    }
-  };
-
-  handleWantedMissing = (body) => {
-    if (body.action === 'updated') {
-      this.props.dispatchUpdateItem({
-        section: 'wanted.missing',
-        updateOnly: true,
-        ...body.resource,
-      });
-    }
-  };
-
-  handleSystemTask = () => {
-    queryClient.invalidateQueries({ queryKey: ['/system/task'] });
-  };
-
-  handleRootfolder = () => {
-    this.props.dispatchFetchRootFolders();
-  };
-
-  handleTag = (body) => {
-    if (body.action === 'sync') {
-      this.props.dispatchFetchTags();
-      this.props.dispatchFetchTagDetails();
-      return;
-    }
-  };
-
-  //
-  // Listeners
-
-  onStartFail = (error) => {
-    console.error('[signalR] failed to connect');
-    console.error(error);
-
-    this.props.dispatchSetAppValue({
-      isConnected: false,
-      isReconnecting: false,
-      isDisconnected: false,
-      isRestarting: false,
-    });
-  };
-
-  onStart = () => {
-    console.debug('[signalR] connected');
-
-    this.props.dispatchSetAppValue({
-      isConnected: true,
-      isReconnecting: false,
-      isDisconnected: false,
-      isRestarting: false,
-    });
-  };
-
-  onReconnecting = () => {
-    this.props.dispatchSetAppValue({ isReconnecting: true });
-  };
-
-  onReconnected = () => {
-    const { dispatchFetchCommands, dispatchSetAppValue } = this.props;
-
-    dispatchSetAppValue({
-      isConnected: true,
-      isReconnecting: false,
-      isDisconnected: false,
-      isRestarting: false,
-    });
-
-    // Repopulate the page (if a repopulator is set) to ensure things
-    // are in sync after reconnecting.
-    queryClient.invalidateQueries({ queryKey: ['/movie/paged'] });
-    queryClient.invalidateQueries({ queryKey: ['/movie/stats'] });
-    dispatchFetchCommands();
-    repopulatePage();
-  };
-
-  onClose = () => {
-    console.debug('[signalR] connection closed');
-  };
-
-  onReceiveMessage = (message) => {
-    console.debug('[signalR] received', message.name, message.body);
-
-    this.handleMessage(message);
-  };
-
-  //
-  // Render
-
-  render() {
-    return null;
-  }
-}
-
-SignalRConnector.propTypes = {
-  isReconnecting: PropTypes.bool.isRequired,
-  isDisconnected: PropTypes.bool.isRequired,
-  isQueuePopulated: PropTypes.bool.isRequired,
-  dispatchFetchCommands: PropTypes.func.isRequired,
-  dispatchUpdateCommand: PropTypes.func.isRequired,
-  dispatchFinishCommand: PropTypes.func.isRequired,
-  dispatchSetAppValue: PropTypes.func.isRequired,
-  dispatchSetVersion: PropTypes.func.isRequired,
-  dispatchUpdate: PropTypes.func.isRequired,
-  dispatchUpdateItemsBatch: PropTypes.func.isRequired,
-  dispatchUpdateItem: PropTypes.func.isRequired,
-  dispatchRemoveItem: PropTypes.func.isRequired,
-  dispatchFetchQualityDefinitions: PropTypes.func.isRequired,
-  dispatchFetchQueue: PropTypes.func.isRequired,
-  dispatchFetchQueueDetails: PropTypes.func.isRequired,
-  dispatchFetchRootFolders: PropTypes.func.isRequired,
-  dispatchFetchTags: PropTypes.func.isRequired,
-  dispatchFetchTagDetails: PropTypes.func.isRequired,
-};
-
-export default connect(
-  createMapStateToProps,
-  mapDispatchToProps
-)(SignalRConnector);
